% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_response.R
\name{normalize_response}
\alias{normalize_response}
\title{Normalize and validate a WCST user response}
\usage{
normalize_response(response)
}
\arguments{
\item{response}{A character string input from the user.}
}
\value{
A cleaned string if valid, or NA if invalid.
}
\description{
Normalize and validate a WCST user response
}
