% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wcst_performance.R
\name{plot_wcst_performance}
\alias{plot_wcst_performance}
\title{Plot WCST session performance}
\usage{
plot_wcst_performance(session_df)
}
\arguments{
\item{session_df}{A data frame from \code{simulate_wcst_session()}.}
}
\value{
A ggplot object showing trial-by-trial performance.
}
\description{
Visualizes trial-by-trial correctness and rule switches.
}
\examples{
res <- simulate_wcst_session(30, seed = 123)
plot_wcst_performance(res)
}
